requirements.txt
nicegui
python-chess
plotly
httpx